array based linked list
An array-based linked list is similar to our Node-based linked list.
However, instead of allocating a single Node at a time, we allocate an
array of Nodes at a time and the Nodes contain the data and a Next
pointer to the next piece of data (or -1 for null). The array contains a
bunch of null nodes at the beginning and these represent the set of free
nodes. Free nodes are all linked together just like a linked list, but
you wouldn’t want to look at the data. When space is needed, we must
check if we have a free node, if we do, put that in our list and fix the
free list. When we remove a node, give it back to the free list. We are
essentially managing two lists.